fix(infra): Manually manage the version of Erlang cluster state (#6087)
Whenever we deploy we risk of having a "split brain" every time
deployment doesn't finalize for all the components. This causes our API
layer (relays, clients, gateways) to see each other but their status
(due to missing presence) is rendered incorrectly in web.

To solve this we will be managing the cluster state manually and
changing it version every time there is a breaking change to the state,
this is more prone to "human errors" but, since we don't change it much,
still better for overall uptime of the portal.