chore: optimise `tunnel_test` debug output (#6088)
When `tunnel_test` fails, it prints the initial state in verbose debug
formatting. Most of the fields in `RefClient` track state _during_ the
runtime of the test and are all empty initially. The same thing applies
to `Host`.

To make this output easier to read and scroll, we ignore some of these
fields in the debug output.