build(deps): use upstream str0m (#5745)
With the adoption of #5080, connlib is now resilient against temporarily
failed connections as they'll be immediately re-established. Thus, we no
longer need any of the patches that we are currently maintaining in our
str0m fork.

The only difference is an adjustment of the ICE timeout parameters but
those can be made configurable in str0m.

Related: https://github.com/algesten/str0m/pull/537.