feat(android): add setDisabledResources FFI (#6166)
Builds on top of  #6164

Part of the effor towards
https://github.com/firezone/firezone/issues/6074

Prepares connlib to call `setDisableResource` from android.

Furthermore, we add a `disablable` parameter for resources which default
to false for now, in the future the portal will set it for the internet
resource, and further in the future it may be used for other resources.

The `disablable` parameter only affect UI.