ci: fix flaky integration test (#6245)
My attempts at fixing the flaky integration test with #6200 failed.
Unfortunately, we need to go back to a fork of str0m to fix this. I am
in discussion with the other maintainers on whether or not we can land
this patch.

Diff to upstream `str0m`:
https://github.com/algesten/str0m/compare/main...firezone:str0m:main
Example of a flaky test run:
https://github.com/firezone/firezone/actions/runs/10328744448/job/28595705615?pr=6237

Related: https://github.com/algesten/str0m/pull/489.