test(connlib): drop DNS queries to CIDR resources for pending connections (#6273)
In #6259, we added a regression test for concurrent DNS queries. A case
that we overlooked is that when DNS servers are defined as CIDR
resources, the queries themselves will act as connection intents and
thus dropped until we have a connection.

In the tests, the connection is only established as part of `advance`.
Thus, if we get multiple concurrent DNS queries to the same server that
is defined as a CIDR resource, we need to drop all future queries.

Fixes: #6283.

---------

Signed-off-by: Thomas Eizinger <thomas@eizinger.io>